Struct tankerkoenig::api::station::StationApi

pub struct StationApi { /* private fields */ }

Struct that holds the current reqwest client of the library and gives access to the station api of the tankerkoenig API.

Implementations

impl StationApi

pub async fn fetch_near( &self, lat: f64, lng: f64, radius: f64, ) -> Result<AreaNearResponse, TankerkoenigError>

Fetch all stations near a given area with some basic informations

Example

use tankerkoenig::Tankerkoenig;
use tankerkoenig::models;

async fn request_near_stations() -> Result<models::AreaNearResponse, tankerkoenig::Error> {
   let tanker = Tankerkoenig::new("your-api-key")?;
   let latitude: f64 = 52.52;
   let longitude: f64 = 13.40;
   let radius: f64 = 10.0;
   let near_stations_res = tanker.station.fetch_near(latitude, longitude, radius).await?;
   Ok(near_stations_res)
}

pub async fn fetch_by_fuel( &self, lat: f64, lng: f64, radius: f64, fuel: Fuel, sort: Sort, ) -> Result<AreaFuelResponse, TankerkoenigError>

Fetch all stations in a radius around the given coordinates that sell a specific kind of Fuel and sort them in a certain order.

Example

use tankerkoenig::Tankerkoenig;
use tankerkoenig::models;

async fn request_fuel_near() -> Result<models::AreaFuelResponse, tankerkoenig::Error> {
   let tanker = Tankerkoenig::new("your-api-key")?;
   let latitude: f64 = 52.52;
   let longitude: f64 = 13.40;
   let radius: f64 = 10.0;
   let fuel = models::Fuel::Diesel;
   let sorting = models::Sort::Distance;
   let stations = tanker.station.fetch_by_fuel(latitude, longitude, radius, fuel, sorting).await?;
   Ok(stations)
}

pub async fn fetch_details<S: AsRef<str>>( &self, id: S, ) -> Result<DetailsResponse, TankerkoenigError>

Fetch informations about a certain station by id.

Example

use tankerkoenig::Tankerkoenig;
use tankerkoenig::models;

async fn request_station() -> Result<models::DetailsResponse, tankerkoenig::Error> {
   let tanker = Tankerkoenig::new("your-api-key")?;
   let station_details = tanker.station.fetch_details("<station_id>").await?;
   Ok(station_details)
}